Create auto-run-test
this needs to be tested on the rpi. intended to be runnable without using a screen, relying on auto-start at power on and showing the status on the display, waiting for the button press to start/stop recording.

only records sensor data when there is valid GPS signals.

needs troubleshooting.